LayoutTests/fast/frames/seamless/seamless-percent-height-expected.txt

Test that the 'initial containing block height' is 0 for seamless child documents.
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "0px"
